requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
idna==3.11
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
drf-spectacular==0.27.2
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
